requirements.txt
python-telegram-bot==21.0.1
requests
python-dotenv
Brotli
pycryptodome
qrcode
Pillow